[quant][refactor tests] split test_workflow_module into test_workflow_ops and test_workflow_module Summary: some tests are used to check the op level numerics of the fake quantize operations
Test Plan:
python test/test_quantization.py

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
